Reorganize the JupyterHub build
Move the spawner and authenticator into separate Python modules
with their own build systems and normal, unpinned dependencies.
Hide the internal modules more effectively and use the new
rubin.nublado.* namespace for both. Unpinning dependencies allows
both modules and the controller to be tested in the same nox session.

Move the infrastructure for the JupyterHub build, including its
pinned dependencies, into a separate hub subdirectory. Only it pins
dependencies, and it no longer needs development dependencies.

This should fix dependency conflicts between the Hub and controller
builds.
