+++ Dockerfile.hub
@@ -17,17 +17,17 @@
 FROM jupyterhub/jupyterhub:4.0.2 as base-image
 
 # Update system packages
-COPY spawner/scripts/install-base-packages.sh .
+COPY hub/scripts/install-base-packages.sh .
 RUN ./install-base-packages.sh && rm install-base-packages.sh
 
 FROM base-image as dependencies-image
 
 # Install system packages only needed for building dependencies.
-COPY spawner/scripts/install-dependency-packages.sh .
+COPY hub/scripts/install-dependency-packages.sh .
 RUN ./install-dependency-packages.sh && rm install-dependency-packages.sh
 
 # Install the dependencies of the extra JupyterHub modules.
-COPY spawner/requirements/main.txt ./requirements.txt
+COPY hub/requirements/main.txt ./requirements.txt
 RUN pip install --quiet --no-cache-dir -r requirements.txt
 
 FROM dependencies-image as runtime-image
@@ -35,7 +35,7 @@ FROM dependencies-image as runtime-image
 # Install the extra JupyterHub modules.
 COPY . /app
 WORKDIR /app
-RUN pip install --no-cache-dir spawner
+RUN pip install --no-cache-dir ./authenticator ./spawner
 
 # Create a non-root user to run JupyterHub. Upstream uses 768 as the UID
 # and GID. This value must be kept in sync with the Phalanx nublado
