+++ authenticator/LICENSE
@@ -1,0 +1,1 @@
+../LICENSE
