+++ authenticator/src/rubin/nublado/authenticator/__init__.py
@@ -1,0 +1,5 @@
+"""Gafaelfawr authenticator for JupyterHub."""
+
+from ._internals import GafaelfawrAuthenticator
+
+__all__ = ["GafaelfawrAuthenticator"]
