--- spawner/src/rsp_restspawner/__init__.py
@@ -1,4 +1,0 @@
-from .auth import GafaelfawrAuthenticator
-from .spawner import RSPRestSpawner
-
-__all__ = ["GafaelfawrAuthenticator", "RSPRestSpawner"]
