+++ spawner/src/rubin/nublado/spawner/__init__.py
@@ -1,0 +1,3 @@
+from ._internals import RSPRestSpawner
+
+__all__ = ["RSPRestSpawner"]
