+++ spawner/tests/support/controller.py
@@ -9,7 +9,7 @@ from datetime import UTC, datetime, timedelta
 import respx
 from httpx import AsyncByteStream, Request, Response
 
-from rsp_restspawner.spawner import LabStatus
+from rubin.nublado.spawner._models import LabStatus
 
 __all__ = [
     "MockLabController",
